Enable GitHub Actions Build
This enables automated builds of this repo via QEMU. Because this is quite expensive there is also a x86_64-based check to see if we need to build at all. Also there is some prep/alignment with https://github.com/brian-maloney/aur-build for potential move of this functionality into that repo someday.